Cleaned up the Parser and added all the Bukkit-Imports.
Deleted Parser main and moved method partly in other classes.

Provided files for all the Bukkit imports I could provide as well as a
Utility to easily create one.
Added a class "DefaultImports" containing all the default ones.

So, as bwfcwalshy freed 3 bit of Ram, I will throw nearly 15,000 Strings
in it xD (Did that before too, to be fair. And this is just because I
can't directly check if the class file exists, but must cache all
possible ones)
Could be done more RAM efficient, but then it would take more CPU and
disk cycles. Not worth it IMO.