Fixed Decompiler not being able to decompile Jars with packaged source
The decompiler will now strip the packaged source before passing it to
fernflower.
Other way would have been ignoring the fernflower warnings, but in this
case fernflower would copy the .java files, not the decompiled class
files.
This means, a user could just modify the .java source files to be clean,
and the .class files to be malicious. This wouldn't be detected by the
Checker, as it would only see the copied .java files, not the
decompiled.
So I chose this overly complicated workaround. Needs some improvement
for the backup function, see the TODO inside.

Also made the public static variables private, provided getters and made
them actual constants, not just be formatted like one.