Added "warn" to logger, enhanced the fix for issue #20
Added a "warn" method to the logger, as it came in handy.

Enhanced the shutdown check to try to find and check the type of the
variable shutdown is invoked onto. Kept ArsenArsen's regex, just added
more handling if it is not
"Bukkit#shutdown()" or "Bukkit#getServer()#shutdown()"
but
"Server server = Bukkit#getServer()"
"server#shutdown()"

This should now be identified correctly. Emphasis on should. I hope this
doesn't introduce any new bugs, having something to test it with would
be nice ;)